Move 6 classes to bletia-core
