+++ bletia-core/src/main/java/info/izumin/android/bletia/core/BleErrorType.java
@@ -1,4 +1,4 @@
-package info.izumin.android.bletia;
+package info.izumin.android.bletia.core;
 
 import android.bluetooth.BluetoothGatt;
